# openapi-facade.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Custom GPT Facade API",
    "version": "2.1.0",
    "description": "BFF (Backend for Frontend) for Custom GPT actions. Provides high-level macros optimized for LLM consumption + unified RPC interface. Timezone: Europe/Prague. All relative time windows (today, lastHour, etc.) are resolved server-side."
  },
  "servers": [
    { "url": "https://api.example.com" }
  ],
...
  "paths": {
...
    "/macros/confirm": {
      "post": {
        "tags": ["confirm"],
        "operationId": "confirmMacroOperation",
        "summary": "Complete pending confirmation",
        "description": "Finalize operation waiting for user confirmation. Pass confirmToken from prior macro call + action (auto-fill, skip, merge, keepBoth). Returns final result.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConfirmRequest" },
              "examples": {
                "merge": {
                  "summary": "Confirm deduplication merge",
                  "value": {
                    "confirmToken": "xyz789...",
                    "action": "merge"
                  }
                },
                "auto_fill": {
                  "summary": "Confirm enrichment auto-fill",
                  "value": {
                    "confirmToken": "abc123def456",
                    "action": "auto-fill"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation completed",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ConfirmResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/macros/confirm/{confirmToken}": {
      "get": {
        "tags": ["confirm"],
        "operationId": "getPendingConfirmationPreview",
        "summary": "Preview pending confirmation",
        "description": "Inspect what a confirmToken contains before confirming. Shows operation type, preview data, and available actions. Helps user decide (auto-fill vs skip, merge vs keepBoth).",
        "parameters": [
          {
            "name": "confirmToken",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Token from prior macro call"
          }
        ],
        "responses": {
          "200": {
            "description": "Confirmation preview",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ConfirmationPreview" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/macros/confirm/{confirmToken}/cancel": {
      "post": {
        "tags": ["confirm"],
        "operationId": "cancelPendingConfirmation",
        "summary": "Cancel pending confirmation",
        "description": "Abandon a waiting confirmation. Clears the confirmToken and reverts operation state.",
        "parameters": [
          {
            "name": "confirmToken",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled",
            "content": {
              "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "message": { "type": "string" } } } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmRequest": {
        "type": "object",
        "properties": {
          "confirmToken": { "type": "string", "description": "From prior macro response." },
          "action": { "type": "string", "enum": ["auto-fill", "skip", "merge", "keepBoth"], "description": "auto-fill: use suggested enrichment. skip: create without. merge: merge duplicates. keepBoth: add both." }
        },
        "required": ["confirmToken", "action"],
        "additionalProperties": false
      },
      "ConfirmResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "data": { "type": "object", "description": "Operation result (event, contact list, etc.)." }
        },
        "required": ["ok"]
      },
      "ConfirmationPreview": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "confirmToken": { "type": "string" },
          "preview": {
            "type": "object",
            "properties": {
              "type": { "type": "string", "enum": ["enrichment", "deduplication"] },
              "operation": { "type": "string" },
              "availableActions": { "type": "array", "items": { "type": "string" } },
              "details": { "type": "object", "description": "Operation-specific preview data." }
            }
          },
          "expiresAt": { "type": "string", "format": "date-time", "description": "Token expiration (usually 10 minutes)." }
        },
        "required": ["ok", "confirmToken", "preview"]
      },
...
    },
    "responses": {
      "ClientError": {
        "description": "4xx client error (invalid request, missing params, enum violation, etc.)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string", "maxLength": 300 },
                "code": { "type": "string", "enum": ["INVALID_PARAM", "BAD_REQUEST", "CONFIRM_SELF_SEND_REQUIRED", "CONFIRMATION_EXPIRED"] },
                "hint": { "type": ["string", "null"], "maxLength": 200 }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "401 - Not authenticated or token expired",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string" }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
      "NotFound": {
        "description": "404 - Resource not found (confirmation token expired, event not found, etc.)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string" }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
...
      "ServerError": {
        "description": "5xx server error (temporary issue, please retry)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string" }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
      "RateLimit429": {
        "description": "Too Many Requests - rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "rate_limit_exceeded" },
                "retryAfterSeconds": { "type": "integer", "minimum": 1 }
              },
              "required": ["error"]
            }
          }
        },
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": { "type": "integer", "minimum": 1 }
          }
        }
      }
    }
  }
}